AI_in_education/Career_Mapping/requirements.txt
# Core Dependencies
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
pymongo==4.6.0

# Development Dependencies
pytest==7.4.2
pytest-cov==4.1.0
black==23.9.1
flake8==6.1.0

# Runtime Dependencies
numpy==1.26.0
pandas==2.1.1
scikit-learn==1.3.0

# Web Server
waitress==2.1.2
gunicorn==21.2.0

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==1.3.0

# Code Quality
pre-commit==3.3.3
mypy==1.5.1

# Environment Management
virtualenv==20.24.5
